edit:postUD:작성자만 UD권한부여

+++ controllers/post/delete.js
@@ -1,3 +1,4 @@
+const mongoose = require("mongoose");
 const { posts } = require("../../mongodb/models");
 const { isAuthorized } = require("../tokenHandle");
 
@@ -11,6 +12,17 @@ module.exports = async (req, res) => {
         message: "Authorization dont exist"
       });
     };
+
+    const accord = await posts.findOne({ 
+      _id: post_id, 
+      "userInfo.user_id": mongoose.Types.ObjectId(tokenBodyData.user_id) 
+    });
+    if(!accord) {
+      return res.status(401).send({
+        data: null,
+        message: "Authorization dont exist"
+      });
+    }
 
     await posts.findByIdAndDelete(post_id)
     .then((result) => {
